# gateway/api/openapi.yaml
openapi: "3.1.0"
info:
  title: "Nebu Admin API"
  version: "1.0.0"
servers:
  - url: "/api/v1"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
    EmptyResponse:
      type: object
      description: Placeholder response; replaced in Epic 6 sub-stories
    DeactivateUserRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 10
    UserStatusResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - user_id
            - status
          properties:
            user_id:
              type: string
            status:
              type: string
    UserListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            type: object
        meta:
          type: object
          required:
            - total
          properties:
            total:
              type: integer
            next_cursor:
              type: string
    UserDetailResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
    AssignUserRoleRequest:
      type: object
      required:
        - role
        - action
      properties:
        role:
          type: string
          enum: [instance_admin, compliance_officer]
        action:
          type: string
          enum: [grant, revoke]
    AssignUserRoleResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - user_id
            - role
            - action
          properties:
            user_id:
              type: string
            role:
              type: string
            action:
              type: string
security:
  - BearerAuth: []
paths:
  /health:
    get:
      operationId: GetHealth
      summary: Health check (unauthenticated)
      security: []
      responses:
        "200":
          description: Health status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /admin/users:
    get:
      operationId: ListAdminUsers
      summary: List users (instance_admin required)
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "400":
          description: Bad request (invalid cursor or limit)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "501":
          description: Not implemented
  /admin/users/{userId}:
    get:
      operationId: GetAdminUser
      summary: Get single user (instance_admin required)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
  /admin/users/{userId}/deactivate:
    post:
      operationId: DeactivateAdminUser
      summary: Deactivate user account (instance_admin required)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeactivateUserRequest"
      responses:
        "200":
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStatusResponse"
        "400":
          description: Bad request (missing/short reason)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "409":
          description: Conflict (already deactivated)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "501":
          description: Not implemented
  /admin/users/{userId}/reactivate:
    post:
      operationId: ReactivateAdminUser
      summary: Reactivate user account (instance_admin required)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User reactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStatusResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "409":
          description: Conflict (irreversible state or already active)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "501":
          description: Not implemented
  /admin/users/{userId}/roles:
    post:
      operationId: AssignAdminUserRole
      summary: Grant or revoke a role override for a user (instance_admin required)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignUserRoleRequest"
      responses:
        "200":
          description: Role assigned or revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignUserRoleResponse"
        "400":
          description: Bad request (invalid role or action)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "403":
          description: Forbidden (self-revoke protection)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "404":
          description: User or role override not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "501":
          description: Not implemented
  /admin/rooms:
    get:
      operationId: ListAdminRooms
      summary: List rooms (placeholder — Story 6.7)
      responses:
        "200":
          description: Room list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "501":
          description: Not implemented
  /admin/config:
    get:
      operationId: GetAdminConfig
      summary: Get server config (placeholder — Story 6.10)
      responses:
        "200":
          description: Config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "501":
          description: Not implemented
  /admin/metrics:
    get:
      operationId: GetAdminMetrics
      summary: Get metrics (placeholder — Story 6.10)
      responses:
        "200":
          description: Metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "501":
          description: Not implemented
  /compliance/access-requests:
    get:
      operationId: ListComplianceAccessRequests
      summary: List compliance access requests (placeholder — Story 6.x)
      responses:
        "200":
          description: Access requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "501":
          description: Not implemented